refactor(visualization): consolidate cc.json read path into loadFile + pin the wire DTO
loadFile now owns cc.json ingestion: moved urlExtractor in, and extracted the
wire-parsing functions (getCCFile, getCCFileAndDecorateFileChecksum) out of the
mixed util/fileHelper into loadFile/util/ccFileHelper. The cross-feature callers
reach them through loadFile's facade (navBar upload, globalSettings url-reset).
util/fileHelper keeps only getSelectedFilesSize (a FileState helper, no wire
format). fileDownloader (export side) stays shared.

Adds a dependency-cruiser rule 'wire-dto-only-in-serialization-boundary' that
confines codeCharta.api.model to the loadFile feature, the navBar gameObjects
importer, and util/fileDownloader — so rendering/state/UI can't couple to the
wire format and the upcoming cc.json 2.0 change stays contained to the converter.

tsc 0 errors, dependency-cruiser 0 errors / 94 warnings, tests green.
