+++ visualization/app/codeCharta/features/globalSettings/components/globalConfigurationDialog/confirmResetMapDialog/confirmResetMapDialog.spec.ts
@@ -8,7 +8,7 @@ import "fake-indexeddb/auto"
 import { getNameDataPair } from "../../../../loadFile/facade"
 import { LoadFileService } from "../../../../../features/loadFile/facade"
 import { LoadInitialFileService, sampleFile1, sampleFile2 } from "../../../../../features/loadFile/facade"
-import { UrlExtractor } from "../../../../../util/urlExtractor"
+import { UrlExtractor } from "../../../../loadFile/facade"
 import * as resetChosenMetricsEffect from "../../../../../state/effects/resetChosenMetrics/resetChosenMetrics.effect"
 import { metricDataSelector } from "../../../../../state/selectors/accumulatedData/metricData/metricData.selector"
 import { setState } from "../../../../../state/store/state.actions"
@@ -18,7 +18,7 @@ import * as indexedDBWriter from "../../../../../util/indexedDB/indexedDBWriter"
 import { ConfirmResetMapDialogComponent } from "./confirmResetMapDialog.component"
 
 jest.mock("../../../../../util/indexedDB/indexedDBWriter")
-jest.mock("../../../../../util/urlExtractor")
+jest.mock("../../../../loadFile/facade")
 jest.mock("../../../../../state/effects/resetChosenMetrics/resetChosenMetrics.effect")
 
 describe("ConfirmResetMapDialogComponent", () => {
