+++ visualization/app/codeCharta/features/loadFile/facade.ts
@@ -1,3 +1,5 @@
 export { LoadFileService } from "./services/loadFile.service"
 export { LoadInitialFileService, sampleFile1, sampleFile2 } from "./services/loadInitialFile.service"
 export { getNameDataPair } from "./util/fileParser"
+export { getCCFile, getCCFileAndDecorateFileChecksum } from "./util/ccFileHelper"
+export { UrlExtractor } from "./util/urlExtractor"
