+++ visualization/app/codeCharta/features/loadFile/services/loadInitialFile.service.spec.ts
@@ -26,10 +26,10 @@ import { ErrorDialogService } from "../../../features/shared/components/errorDia
 import { getNameDataPair } from "../util/fileParser"
 import { LoadFileService } from "./loadFile.service"
 import { LoadInitialFileService, sampleFile1, sampleFile2 } from "./loadInitialFile.service"
-import { UrlExtractor } from "../../../util/urlExtractor"
+import { UrlExtractor } from "../util/urlExtractor"
 import { setCurrentFilesAreSampleFiles } from "../../../state/store/appStatus/currentFilesAreSampleFiles/currentFilesAreSampleFiles.actions"
 
-jest.mock("../../../util/urlExtractor")
+jest.mock("../util/urlExtractor")
 jest.mock("../../../model/files/files.helper")
 jest.mock("../../../util/indexedDB/indexedDBWriter")
