+++ visualization/app/codeCharta/features/loadFile/util/fileParser.ts
@@ -2,7 +2,7 @@ import { ExportBlacklistItem, ExportCCFile } from "../../../codeCharta.api.model
 import { CCFile } from "../../../codeCharta.model"
 import { NameDataPair } from "../../../codeCharta.api.model"
 import { FileSelectionState, FileState } from "../../../model/files/files"
-import { getCCFile } from "../../../util/fileHelper"
+import { getCCFile } from "./ccFileHelper"
 import { CCFileValidationResult as FileValidationResult, checkErrors, checkWarnings, removeAuthorsAttributes } from "./fileValidator"
 import { NodeDecorator } from "../../../util/nodeDecorator"
